+++ src/impls.rs
@@ -1,4 +1,4 @@
-// Copyright 2019-2021 Parity Technologies (UK) Ltd.
+// Copyright 2019-2022 Parity Technologies (UK) Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -41,7 +41,6 @@ use crate::{
     TypeDefArray,
     TypeDefCompact,
     TypeDefPrimitive,
-    TypeDefRange,
     TypeDefSequence,
     TypeDefTuple,
     TypeInfo,
@@ -359,7 +358,14 @@ where
 {
     type Identity = Self;
     fn type_info() -> Type {
-        TypeDefRange::new::<Idx>(false).into()
+        Type::builder()
+            .path(Path::prelude("Range"))
+            .type_params(type_params![Idx])
+            .composite(
+                Fields::named()
+                    .field(|f| f.name("start").ty::<Idx>().type_name("Idx"))
+                    .field(|f| f.name("end").ty::<Idx>().type_name("Idx")),
+            )
     }
 }
 
@@ -369,7 +375,14 @@ where
 {
     type Identity = Self;
     fn type_info() -> Type {
-        TypeDefRange::new::<Idx>(true).into()
+        Type::builder()
+            .path(Path::prelude("RangeInclusive"))
+            .type_params(type_params![Idx])
+            .composite(
+                Fields::named()
+                    .field(|f| f.name("start").ty::<Idx>().type_name("Idx"))
+                    .field(|f| f.name("end").ty::<Idx>().type_name("Idx")),
+            )
     }
 }
 
@@ -377,15 +390,15 @@ where
 mod bit_vec {
     use super::*;
 
-    impl<O, T> TypeInfo for bitvec::vec::BitVec<O, T>
+    impl<T, O> TypeInfo for bitvec::vec::BitVec<T, O>
     where
-        O: bitvec::order::BitOrder + TypeInfo + 'static,
         T: bitvec::store::BitStore + TypeInfo + 'static,
+        O: bitvec::order::BitOrder + TypeInfo + 'static,
     {
         type Identity = Self;
 
         fn type_info() -> Type {
-            crate::TypeDefBitSequence::new::<O, T>().into()
+            crate::TypeDefBitSequence::new::<T, O>().into()
         }
     }