+++ test_suite/tests/ui/fail_with_invalid_scale_info_attrs.rs
@@ -3,11 +3,10 @@ use scale_info::TypeInfo;
 use scale::Encode;
 
 #[derive(TypeInfo, Encode)]
-#[scale_info(capture_docs = "invalid")]
-/// Docs
-struct InvalidDocsCapture {
-    /// Docs
+#[scale_info(foo)]
+struct InvalidKeywordInScaleInfoAttr {
     a: u8,
+    b: u16,
 }
 
 fn main() {}